# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Public Atlas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/institutions": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "List Institutions",
        "description": "A page of institutions. `q` matches a name or an alias; `place_id` admits the place and\nevery place under it.",
        "operationId": "graph-list_institutions",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Code"
            }
          },
          {
            "name": "place_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place Id"
            }
          },
          {
            "name": "administrative_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Administrative Level"
            }
          },
          {
            "name": "institution_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Institution Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EntityStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "parent_institution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Institution Id"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "name",
                "institution_type",
                "status",
                "created_at",
                "place"
              ],
              "type": "string",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_InstitutionOutput_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/institutions/{institution_id}": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Read Institution",
        "description": "The institution with its aliases, its place and the places above, its parent and the\nbodies under it, every homepage claim, its sources, and every quote for any of them with a\nlink to the stored copy it was found on.",
        "operationId": "graph-read_institution",
        "parameters": [
          {
            "name": "institution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Institution Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/places": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "List Places",
        "description": "A page of places by name; `q` matches a name or an alias.",
        "operationId": "graph-list_places",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "country_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Z]{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Code"
            }
          },
          {
            "name": "administrative_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Administrative Level"
            }
          },
          {
            "name": "parent_place_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Place Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PlaceOutput_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/places/{place_id}": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Read Place",
        "description": "The place with the places above it and its government.",
        "operationId": "graph-read_place",
        "parameters": [
          {
            "name": "place_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Place Id"
            }
          },
          {
            "name": "x-database",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`main` (the default) or `eval`: which database the request reads.",
              "title": "X-Database"
            },
            "description": "`main` (the default) or `eval`: which database the request reads."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnteredBy": {
        "type": "string",
        "enum": [
          "manual",
          "script",
          "agent"
        ],
        "title": "EnteredBy",
        "description": "How a row got here. On every entity and every evidence row."
      },
...
      "EntityStatus": {
        "type": "string",
        "enum": [
          "candidate",
          "verified",
          "rejected",
          "needs_review"
        ],
        "title": "EntityStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentifierOutput": {
        "properties": {
          "scheme": {
            "type": "string",
            "title": "Scheme"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "scheme",
          "value"
        ],
        "title": "IdentifierOutput"
      },
      "InstitutionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "institution_type": {
            "type": "string",
            "title": "Institution Type"
          },
          "suggested_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Type"
          },
          "status": {
            "$ref": "#/components/schemas/EntityStatus"
          },
          "entered_by": {
            "$ref": "#/components/schemas/EnteredBy"
          },
          "place": {
            "$ref": "#/components/schemas/PlaceRef"
          },
          "parent_institution_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Institution Id"
          },
          "procurement_handled_by": {
            "$ref": "#/components/schemas/ProcurementHandledBy"
          },
          "homepage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Id"
          },
          "homepage_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/AliasOutput"
            },
            "type": "array",
            "title": "Aliases"
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/IdentifierOutput"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricOutput"
            },
            "type": "array",
            "title": "Metrics"
          },
          "places": {
            "items": {
              "$ref": "#/components/schemas/PlaceRef"
            },
            "type": "array",
            "title": "Places"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstitutionRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/InstitutionRef"
            },
            "type": "array",
            "title": "Children"
          },
          "served_places": {
            "items": {
              "$ref": "#/components/schemas/PlaceRef"
            },
            "type": "array",
            "title": "Served Places"
          },
          "homepages": {
            "items": {
              "$ref": "#/components/schemas/HomepageOutput"
            },
            "type": "array",
            "title": "Homepages"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOutput"
            },
            "type": "array",
            "title": "Sources"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOutput"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "institution_type",
          "suggested_type",
          "status",
          "entered_by",
          "place",
          "parent_institution_id",
          "procurement_handled_by",
          "homepage_id",
          "homepage_url",
          "created_at",
          "aliases",
          "identifiers",
          "metrics",
          "places",
          "parent",
          "children",
          "served_places",
          "homepages",
          "sources",
          "evidence"
        ],
        "title": "InstitutionDetail"
      },
      "InstitutionOutput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "institution_type": {
            "type": "string",
            "title": "Institution Type"
          },
          "suggested_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Type"
          },
          "status": {
            "$ref": "#/components/schemas/EntityStatus"
          },
          "entered_by": {
            "$ref": "#/components/schemas/EnteredBy"
          },
          "place": {
            "$ref": "#/components/schemas/PlaceRef"
          },
          "parent_institution_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Institution Id"
          },
          "procurement_handled_by": {
            "$ref": "#/components/schemas/ProcurementHandledBy"
          },
          "homepage_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Id"
          },
          "homepage_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "institution_type",
          "suggested_type",
          "status",
          "entered_by",
          "place",
          "parent_institution_id",
          "procurement_handled_by",
          "homepage_id",
          "homepage_url",
          "created_at"
        ],
        "title": "InstitutionOutput",
        "description": "An institution as the table lists it."
      },
...
      "Page_InstitutionOutput_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstitutionOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[InstitutionOutput]"
      },
      "Page_PlaceOutput_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaceOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PlaceOutput]"
      },
...
      "PlaceDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "administrative_level": {
            "type": "string",
            "title": "Administrative Level"
          },
          "parent_place_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Place Id"
          },
          "government_institution_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Government Institution Id"
          },
          "status": {
            "$ref": "#/components/schemas/EntityStatus"
          },
          "entered_by": {
            "$ref": "#/components/schemas/EnteredBy"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/PlaceRef"
            },
            "type": "array",
            "title": "Parents"
          },
          "government": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstitutionRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "country_code",
          "administrative_level",
          "parent_place_id",
          "government_institution_id",
          "status",
          "entered_by",
          "created_at",
          "parents",
          "government"
        ],
        "title": "PlaceDetail"
      },
...
      "PlaceRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "administrative_level": {
            "type": "string",
            "title": "Administrative Level"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "administrative_level"
        ],
        "title": "PlaceRef"
      },
      "ProcurementHandledBy": {
        "type": "string",
        "enum": [
          "self",
          "parent"
        ],
        "title": "ProcurementHandledBy",
        "description": "Whether an institution buys on its own account or its parent buys for it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}